RUE-861: extract source loading orchestration
